% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CancerSEA}
\alias{CancerSEA}
\title{Curated CancerSEA gene sets}
\format{
A list with 14 tibble. And each tibble with columns ensembl_gene_id and symbol.
}
\source{
\url{http://biocc.hrbmu.edu.cn/CancerSEA/goDownload}
}
\usage{
data(CancerSEA)
}
\description{
A curated cancer functional state genes lists from CancerSEA.
}
\keyword{datasets}
