% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hallmark}
\alias{Hallmark}
\title{Cancer hallmark gene sets}
\format{
A list with 50 tibble. And each tibble with columns ensembl_gene_id and symbol.
}
\source{
\url{http://www.gsea-msigdb.org/gsea/msigdb/index.jsp}
}
\usage{
data(Hallmark)
}
\description{
Cancer hallmark gene lists getted using \code{msigdbr} package from MsigDB.
}
\keyword{datasets}
