% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Immport}
\alias{Immport}
\title{Curated immune gene sets}
\format{
A list with 17 tibble. And each tibble with columns ensembl_gene_id and symbol.
}
\source{
\url{https://www.immport.org/}
}
\usage{
data(Immport)
}
\description{
A curated immune gene lists from Immport.
}
\keyword{datasets}
