% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_pathway.R
\name{search_genesets}
\alias{search_genesets}
\title{search gene sets in this package}
\usage{
search_genesets(keywords = NULL, ...)
}
\arguments{
\item{keywords}{the pattern argument in \code{\link[base:grep]{base::grep()}}}

\item{...}{the arguments in \link{grep} function}
}
\description{
Search the genesets in current package.
}
